Rename ListCommitsWithNotes and make method private
This solves two open TODOs and reduces the API surface of the notes
package.
